Fixed compiler warnings

+++ raytracer/samplers/JitterGaussian.cpp
@@ -24,8 +24,8 @@ Ray* JitterGaussian::get_rays(size_t px, size_t py) const {
   std::uniform_real_distribution<double> dist(-step, step);
 
   double totalWeight = 0;
-  for (int y = 0; y < degree; ++y) {
-    for (int x = 0; x < degree; ++x) {
+  for (int y = 0; y < static_cast<int>(degree); ++y) {
+    for (int x = 0; x < static_cast<int>(degree); ++x) {
       double xOffset =
           (-static_cast<int>(degree) + 2 * x + 1) * step + dist(generator);
       double yOffset =
